Add files via upload
Your task is to prepare an analytical report for the HR Department. Based on the analysis, make recommendations for the HR Department on the recruitment strategy, as well as on how to interact with existing employees.

The database contains a set of tables with data about employees of a fictional company. Make an overview of the company's staff. Create a set of research items and then test them against the data. All Analytics must be performed using SQL. The data can then be visualized, but the final dataframes for graphs must also be prepared using SQL.


Your task is to prepare an analytical response for the SMM Department of Skillbox.
The object of analysis is the public Skillbox Vkontakte (https://vk.com/skillbox_education).

Connect to the VK API and upload posts from the Skillbox public wall for the period of interest (determine it yourself and justify it). Analyze the impact of various factors (for example, the time of publication) on user engagement (the number of likes, comments, and votes in surveys). Make Analytics for categories (examples of categories: design battle, management game) that are in the public. You can use regular expressions to select the posts you want. Make a list of recommendations for the SMM Department based on the results of the analysis.